Various bugfixes and extending interfaces to use longs when perf isn't a concern.

+++ java/ham_fisted/RangeList.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 
 public interface RangeList {
-  public void fillRange(int startidx, int endidx, Object v);
-  public void fillRange(int startidx, List v);
-  public void removeRange(int startidx, int endidx);
+  public void fillRange(long startidx, long endidx, Object v);
+  public void fillRange(long startidx, List v);
+  public void removeRange(long startidx, long endidx);
 }
